complex code to compute the automatic icon size
--whenever the layout is not hovered but its size
is changed (probably additions, removals from tasks
or applets) an update in icon size is checked.
The update takes into consideration if this was
an addition or a removal and first tries to
predict the impact if an icon size occured.
if the new icon size predicts a correct sized layout
then it is enabled