respect first Now Dock Panel differently
--first Now Dock Panel behaves totally as
Qt::Tool window (this provides us with very
beautiful located pop ups for all the plasmoids)
and all the rest Now Dock Panels (e.g. the second,
third etc.) behave in a combination of
Dock window and an Tool one for some of the dock's
states. This was chosen because if there are
more than one Qt::Tool windows then all the rest
except the first are always on top